requirements.txt
apex==0.9.10dev
g2p_en==2.1.0
inflect==5.6.2
librosa==0.9.2
matplotlib==3.5.2
numpy==1.22.4
pydub==0.25.1
pypinyin==0.39.0
pyworld==0.2.10
PyYAML==6.0
scikit_learn==1.1.1
scipy==1.8.1
tensor2tensor==1.15.7
tensorboardX==2.5.1
tensorflow==2.9.1
tgt==1.4.4
torch==1.12.0
tqdm==4.46.1
Unidecode==1.1.1
noisereduce==2.0.1

# Python 3.10
# apt-get install python-tk
# pip3 install python3-tk
# ffnpeg (jouer les sons) sur ubuntu linux